add create/delete/modify items methods with controllers and tests

+++ eCommerce-backend/src/main/java/com/app/domain/item/services/ItemService.java
@@ -2,6 +2,8 @@ package com.app.domain.item.services;
 
 import com.app.domain.item.dtos.ItemDetailedDTO;
 import com.app.domain.item.dtos.ItemSummaryDTO;
+import com.app.domain.item.dtos.requests.ModifiedItemRequest;
+import com.app.domain.item.dtos.requests.NewItemRequest;
 import com.app.domain.item.entities.Category;
 import com.app.domain.item.entities.Item;
 import com.app.domain.item.exceptions.ItemNotFoundException;
@@ -9,13 +11,15 @@ import com.app.domain.item.mappers.ItemMapper;
 import com.app.domain.item.repositories.ItemRepository;
 import com.app.domain.member.entities.Member;
 import com.app.domain.member.services.MemberService;
+import com.app.global.exceptions.ForbiddenException;
+import com.app.global.utils.AuthUtil;
+import com.app.global.vos.Media;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.Set;
-import java.util.UUID;
+import java.util.*;
 
 @Service
 @Transactional(readOnly = true)
@@ -32,13 +36,61 @@ public class ItemService {
     }
 
     @Transactional
+    public ItemSummaryDTO create(NewItemRequest request) {
+        Member seller = AuthUtil.getAuthenticated();
+        List<Media> mediaList = Collections.emptyList();
+//        List<Media> mediaList = AwsS3Service.upload(request.media()); // TODO
+        Category category = request.categoryId() != null ? categoryService.findById(request.categoryId()) : null;
+        Item item = ItemMapper.toItem(request, category, seller, mediaList);
+        return save(item);
+    }
+
+    @Transactional
+    public ItemSummaryDTO modify(UUID itemId, ModifiedItemRequest request) {
+        Member modifier = AuthUtil.getAuthenticated();
+        Item item = findById(itemId);
+        Member originalSeller = item.getSeller();
+        if (!modifier.equals(originalSeller)) {
+            throw new ForbiddenException();
+        }
+
+        updateItemValues(item, request);
+
+        return save(item);
+    }
+
+    private void updateItemValues(Item item, ModifiedItemRequest request) {
+        if (request.title() != null) {
+            item.setTitle(request.title());
+        }
+        if (request.price() != null) {
+            item.setPrice(request.price());
+        }
+        if (request.description() != null) {
+            item.setDescription(request.description());
+        }
+        if (request.media() != null) {
+            item.setMediaList(Collections.emptyList()); // TODO
+        }
+        if (request.categoryId() != null) {
+            Category category = categoryService.findById(request.categoryId());
+            item.setCategory(category);
+        }
+    }
+
+    @Transactional
     public ItemSummaryDTO save(Item item) {
         return ItemMapper.toItemSummaryDTO(itemRepository.save(item));
     }
 
     @Transactional
     public void deleteById(UUID id) {
+        Member member = AuthUtil.getAuthenticated();
         Item item = findById(id);
+        Member originalSeller = item.getSeller();
+        if (!member.equals(originalSeller)) {
+            throw new ForbiddenException();
+        }
         itemRepository.delete(item);
     }
